add copyright notice and rename module
